codina > codinalte
This change is required due to there is a device
named codina.
Samsung was weird and named our device codina as well but
with an extra 3 letter extension.
Name this codinalte same as jflte and d2lte.
This is a full device tree for 3 devices
the blob installer will find the baseband and choose the correct
naming to extract device specific blobs as well as build.prop
modding the device name.